comments from Greg
